Displaced stepping on x86 over relative branches/jumps as is doesn't work nicelly with SSS.  Currently, we copy the branch/jump to the jump pad, single-step it, and then adjust the relative offset after the fact.  The problem is that with SSS, the branch/jump destination relative to the scratch pad could well be an innacessible address, one where we can't put a SSS bkpt at.
So rework the relative branch/jump support to be like ARM's.  IOW,
compute the branch destination, have the inferior execute a nop (so it
still has a chance to dequeue signals, etc.), SSS single-step that
(which will always land in the scratch pad), and then simply adjust
the PC manually to the original branch destination.