Merge handling of TARGET_WAITKIND_EXITED and TARGET_WAITKIND_SIGNALLED into a single switch case.
I was experimenting with alternative fixes to a bug, and noticing I
was always needing to do the same changes twice, once for each of the
TARGET_WAITKIND_EXITED and TARGET_WAITKIND_SIGNALLED paths.  Looking
at the code, one realizes that most of the handling in both code paths
is actually the same.  This unifies the code paths to reduce and avoid
further duplification.

The fix for the bug I ended up not touching this, but I still think
it's worth it to merge these.  I will need to touch this code on
another series...

gdb/
2012-10-25  Pedro Alves  <palves@redhat.com>

	* infrun.c (handle_inferior_event): Merge handling of
	TARGET_WAITKIND_EXITED and TARGET_WAITKIND_SIGNALLED into a single
	switch case.