foll-vfork.exp: Expose more testing on non-HP/UX targets
Well, exposes more to Linux, since we have at the top:

if {![istarget "hppa*-hp-hpux*"] && ![istarget "*-linux*"]} then {
    continue
}

The vfork_and_exec_child_follow_through_step procedure is about
following the child over a "next".  This is currently disabled on non
HP/UX targets, but it's a useful test to run.  AFAICS from the test,
HP/UX can't debug the vfork child until it execs or exits.  The Linux
kernel can, so the test is adjusted to expect the ideal behavior out
of non HP/UX targets (and we get that right on Linux).

gdb/testsuite/
2012-10-25  Pedro Alves  <palves@redhat.com>

	* gdb.base/foll-vfork.exp
	(vfork_and_exec_child_follow_through_step): Don't skip on
	non-HP/UX targets.  Expect the next to only step one line on
	non-HP/UX targets, rather than stopping only after the exec.