foll-vfork.exp: Also test following vfork when the child exits.
foll-vfork.exp currently has tests for both following the vfork
parent, and the vfork child.  However, the vfork shared region is done
with when either the child execs, or exits, and the tests only test
the exec variant.  These events are handled with similar yet
not-quite-the-same code paths in GDB, so it's useful to add that
vector to the test matrix as well.  (gdb/14766 is a case of a bug on
the exit path only.  gdb/14762 is exposed by making the tests cover
the whole testing matrix.)  The patch implements that by adding a new
test source based on foll-vfork.c that has been adjusted to have the
child _exit instead of exec.  The follow-parent tests can all run with
either the exit or exec variant, using the same tcl procedure even
(with_test_prefix is used to distinguish the variants in gdb.sum).
Some of the follow-child tests can too, while others cannot -- the
one's that can't get special tcl procedures for each exec/exit
variant.

(I had originally developed this as a separate new .exp test file,
based on foll-vfork.exp, until I decided that it was better to merge
them.)

(This kfails both gdb/14762 and gdb/14766.  The gdb/14762 kfail is
actually masked by gdb/14766.  The next patch fixes gdb/14766, thus
exposing the gdb/14762 kfail...)

gdb/testsuite/
2012-10-25  Pedro Alves  <palves@redhat.com>

	* gdb.base/foll-vfork-exit.c: New file.
	* gdb.base/foll-vfork.exp (top level): New file-describing
	comment.
	(vfork_child_follow_to_exit): New procedure.
	(tcatch_vfork_then_child_follow): Rename as ...
	(tcatch_vfork_then_child_follow_exec): ... this.
	(tcatch_vfork_then_child_follow_exit): New procedure.
	(do_vfork_and_follow_parent_tests): New procedure, factored out
	from do_vfork_and_exec_tests.
	(do_vfork_and_follow_child_tests_exec): Ditto.
	(do_vfork_and_exec_tests): Delete.
	(do_vfork_and_follow_child_tests_exit): New procedure.
	(top level): Run tests with both the program that has the vfork
	child execing, and the program has the vfork child exiting.