Fix PR gdb/14766 - vfork child exit with "set follow-fork-mode child" broken.
If "set detach-on-fork" is off (which it is by default), and we follow
the vfork child, we don't actually detach from the parent at vfork
time, but instead hold on to it until the child is done with the
shared memory region.  IOW, until the child execs or exits.  There's a
single function that handles all that work for both child execs and
exits, handle_vfork_child_exec_or_exit.  When that function goes about
detaching the parent, it tries to preserve the selected inferior (the
child), but while doing so, it is also trying to restore the selected
thread (as pointed at by inferior_ptid).  That makes sense for execs,
but not so for exits.  For exits, inferior_ptid, which has been
switched to the event ptid, does not point at a valid thread anymore.
Instead it just points at the process's pid.  The result is:

(gdb) r
Starting program:
/home/pedro/gdb/mygit/build/gdb/testsuite/gdb.base/foll-vfork-exit
[New process 4673]
I'm the child!
I'm the proud parent of child #4673!
Couldn't get registers: No such process.
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
Couldn't get registers: No such process.
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
Couldn't get registers: No such process.
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
(gdb)

The fix is not try to restore the thread, but just the inferior.

The bit in tcatch_vfork_then_child_follow_exit that kfails this PR is
just removed, as there's a specific test for this bug in the file.  It
was put there originally as this bug masked the gdb/14762 bug, and
thus the related gdb/14762 kfail in
tcatch_vfork_then_child_follow_exit was not reached.

gdb/
2012-10-25  Pedro Alves  <palves@redhat.com>

	PR gdb/14766

	* infrun.c (handle_inferior_event)
	<TARGET_WAITKIND_EXITED/TARGET_WAITKIND_SIGNALLED>: Switch to
	null_ptid before handling a vfork child exec or exit.  Switch to
	the event ptid afterwards.

gdb/testsuite/
2012-10-25  Pedro Alves  <palves@redhat.com>

	PR gdb/14766

	* gdb.base/foll-vfork.exp (vfork_child_follow_to_exit): Remove
	setup_kfail.
	(tcatch_vfork_then_child_follow_exit): No longer expect "Couldn't
	get registers".