Eliminate all thread_hop_needed logic.  It's wrong in case a breakpoint for the wrong thread hits just while the step is finishing a step.  Let the common breakpoint-for-wrong-thread handling code (within bpstat_what and friends) handle this.
Special care is still needed to handle the case of PC of the thread we
were trying to single-step having changed.  We can't just keep_going
and re-step it, as in that case we can over-step the thread (it was
already done, but we'd ask it to step even further).  That's now
handled in the switch_back_to_stepped_thread change.  As bonus, we're
now using a technique that doesn't lose signals, unlike the old code.

Tested on x86_64 Fedora 17, against pristine mainline, and against a
branch that implements software single-step on x86.