Fix for even more missed events; eliminate thread-hop code.
Even with deferred_step_ptid out of the way, GDB can still lose
watchpoints on software single-step targets.

Here in the thread-hop code:

	  if (!ptid_equal (singlestep_ptid, ecs->ptid)
	      && in_thread_list (singlestep_ptid))
	    {
	      /* If the PC of the thread we were trying to single-step
		 has changed, discard this event (which we were going
		 to ignore anyway), and pretend we saw that thread
		 trap.  This prevents us continuously moving the
		 single-step breakpoint forward, one instruction at a
		 time.  If the PC has changed, then the thread we were
		 trying to single-step has trapped or been signalled,
		 but the event has not been reported to GDB yet.

		 There might be some cases where this loses signal
		 information, if a signal has arrived at exactly the
		 same time that the PC changed, but this is the best
		 we can do with the information available.  Perhaps we
		 should arrange to report all events for all threads
		 when they stop, or to re-poll the remote looking for
		 this particular thread (i.e. temporarily enable
		 schedlock).  */

	     CORE_ADDR new_singlestep_pc
	       = regcache_read_pc (get_thread_regcache (singlestep_ptid));

	     if (new_singlestep_pc != singlestep_pc)
	       {
		 enum gdb_signal stop_signal;

		 if (debug_infrun)
		   fprintf_unfiltered (gdb_stdlog, "infrun: unexpected thread,"
				       " but expected thread advanced also\n");

		 /* The current context still belongs to
		    singlestep_ptid.  Don't swap here, since that's
		    the context we want to use.  Just fudge our
		    state and continue.  */
                 stop_signal = ecs->event_thread->suspend.stop_signal;
                 ecs->event_thread->suspend.stop_signal = GDB_SIGNAL_0;
                 ecs->ptid = singlestep_ptid;
                 ecs->event_thread = find_thread_ptid (ecs->ptid);
                 ecs->event_thread->suspend.stop_signal = stop_signal;
                 stop_pc = new_singlestep_pc;
               }
             else
	       {
		 if (debug_infrun)
		   fprintf_unfiltered (gdb_stdlog,
				       "infrun: unexpected thread\n");

		 thread_hop_needed = 1;
		 stepping_past_singlestep_breakpoint = 1;
		 saved_singlestep_ptid = singlestep_ptid;
	       }
	    }
	}

Note we either end up with thread_hop_needed, ignoring the watchpoint
SIGTRAP, or switch to the stepping thread, again ignoring that the
SIGTRAP could be for some other event.

So the fix is similar to the deferred_step_ptid fix -- defer the
thread hop to _after_ the SIGTRAP had a change of passing through the
regular bpstat handling.  If the wrong thread hits a breakpoint, we'll
just end up with BPSTAT_WHAT_SINGLE, and if nothing causes a stop,
keep_going starts a step-over.

But in order to do that, we need to make bpstat aware of single-step
breakpoints.  So this patch starts convering the single-step
breakpoints to real breakpoints.  It doesn't put them in the regular
breakpoint chain yet though.  More changes would be needed for that.

The stepping_past_singlestep_breakpoint mechanism is really not
necessary -- just setting the thread to step over a breakpoint with
thread->stepping_over_breakpoint=1 has the exact same effect.

Special care is still needed to handle the case of PC of the thread we
were trying to single-step having changed, like in the old code.  We
can't just keep_going and re-step it, as in that case we can over-step
the thread (if it was already done with the step, but hsn't reported
it yet, we'd ask it to step even further).  That's now handled in
switch_back_to_stepped_thread.  As bonus, we're now using a technique
that doesn't lose signals, unlike the old code -- insert a breakpoint
at PC, and resume, which either reports the breakpoint immediately, or
any pending signal.

Tested on x86_64 Fedora 17, against pristine mainline, and against a
branch that implements software single-step on x86.

gdb/
2014-02-21  Pedro Alves  <palves@redhat.com>

	* breakpoint.c (single_step_breakpoints): Change type to struct
	breakpoint array.
	(single_step_gdbarch): Delete.
	(bpstat_stop_status): Handle single-step breakpoints.
	(bpstat_what, bptype_string, print_one_breakpoint_location)
	(init_bp_location): Handle bp_single_step.
	(init_momentary_breakpoint): New, factored out from ...
	(set_momentary_breakpoint): ... this.  Rewrite.
	(init_momentary_breakpoint_at_pc): New, factored out from ...
	(set_momentary_breakpoint_at_pc): ... this.  Rewrite.
	(insert_single_step_breakpoint, remove_single_step_breakpoints)
	(cancel_single_step_breakpoints, detach_single_step_breakpoints):
	Rewrite/adjust to insert/remove real breakpoints instead of using
	deprecated_insert_raw_breakpoint and
	deprecated_remove_raw_breakpoint.
	(single_step_breakpoint_inserted_here_p): Adjust.
	* breakpoint.h (enum bptype) <bp_single_step>: New breakpoint
	type.
	* infrun.c (saved_singlestep_ptid)
	(stepping_past_singlestep_breakpoint): Delete.
	(resume): Remove stepping_past_singlestep_breakpoint handling.
	(proceed): Store the prev_pc of the stepping thread too.
	(init_wait_for_inferior): Adjust.  Clear singlestep_ptid and
	singlestep_pc.
	(enum infwait_states): Delete infwait_thread_hop_state.
	(handle_inferior_event): Adjust.
	(handle_signal_stop): Delete stepping_past_singlestep_breakpoint
	handling and the thread-hop code.  Defer removing single-step
	breakpoints till after bpstat handling.
	(switch_back_to_stepped_thread): Handle the case of the stepped
	thread having advanced already.